+++ src/com/intellij/idea/plugin/hybris/settings/HybrisProjectSettingsComponent.kt
@@ -33,8 +33,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.util.text.VersionComparatorUtil
 import com.intellij.util.xmlb.XmlSerializerUtil
 
-@State(name = "HybrisProjectSettings")
-@Storage(value = STORAGE_HYBRIS_PROJECT_SETTINGS, roamingType = RoamingType.DISABLED)
+@State(name = "HybrisProjectSettings", storages = [Storage(STORAGE_HYBRIS_PROJECT_SETTINGS, roamingType = RoamingType.DISABLED)])
 @Service
 class HybrisProjectSettingsComponent : PersistentStateComponent<HybrisProjectSettings> {
     private val hybrisProjectSettings = HybrisProjectSettings()